[#32] [Add] 최단 경로 탐색 캐싱 적용
