+++ logistic-service/src/main/java/com/sparta/blackyolk/logistic_service/hubroute/application/usecase/HubRoutePathUseCase.java
@@ -3,5 +3,5 @@ package com.sparta.blackyolk.logistic_service.hubroute.application.usecase;
 import com.sparta.blackyolk.logistic_service.hubroute.framework.web.dto.HubRoutePathResponse;
 
 public interface HubRoutePathUseCase {
-    HubRoutePathResponse getShortestPath(String departure, String arrival);
+    HubRoutePathResponse getShortestPath(String departure, String arrival, String currentTimeSlot);
 }
